Work on tests

+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/expunge/ExpungeEverythingService.java
@@ -318,10 +318,12 @@ public class ExpungeEverythingService implements IExpungeEverythingService {
 						nativeQuery.setMaxResults(800);
 						List pids = nativeQuery.getResultList();
 
-						nativeQuery = myEntityManager.createQuery("DELETE FROM " + theEntityType.getSimpleName()
+						if (!pids.isEmpty()) {
+							nativeQuery = myEntityManager.createQuery("DELETE FROM " + theEntityType.getSimpleName()
 								+ " WHERE " + idProperty + " IN (:pids)");
-						nativeQuery.setParameter("pids", pids);
-						nativeQuery.executeUpdate();
+							nativeQuery.setParameter("pids", pids);
+							nativeQuery.executeUpdate();
+						}
 						return pids.size();
 					});
